% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpvca.R
\name{PlotPVCA}
\alias{PlotPVCA}
\title{A function for plotting a result from the PVCA function}
\usage{
PlotPVCA(pvca.res, title)
}
\arguments{
\item{pvca.res}{The output from the 'pvca' function}

\item{title}{The title of the bar plot}
}
\value{
p A ggplot2 object
}
\description{
This function takes as input a proportion vector calculated from the 'pvca' function and generate a bar plot.
}
