% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvca.R
\name{PVCA}
\alias{PVCA}
\title{A function for principal variance component analysis}
\usage{
PVCA(counts, meta, threshold, inter)
}
\arguments{
\item{counts}{The Normalized(e.g. TMM)/ log-transformed reads count matrix from sequencing data (row:gene/feature, col:sample)}

\item{meta}{The Meta data matrix containing predictor variables (row:sample, col:predictor)}

\item{threshold}{The proportion of the variation in read counts explained by top k PCs. This value determines the number of top PCs to be used in pvca.}

\item{inter}{TRUE/FALSE - include/do not include pairwise interactions of predictors}
}
\value{
std.prop.val The vector of proportions of variation explained by each predictor.
}
\description{
The function is written based on the 'pvcaBatchAssess' function of the PVCA R package 
and slightly changed to make it more efficient and flexible for sequencing read counts data. 
(http://watson.nci.nih.gov/bioc_mirror/packages/release/bioc/manuals/pvca/man/pvca.pdf)
}
